拍摄页面优化

+++ app/src/main/java/com/flyingeffects/com/ui/interfaces/view/FUBeautyMvpView.java
@@ -2,7 +2,7 @@ package com.flyingeffects.com.ui.interfaces.view;
 
 public interface FUBeautyMvpView {
 
-    void showCountDown(int num,int countDownStatus,float progress);
+    void showCountDown(float num,int countDownStatus,float progress);
 
     void nowChooseRecordIsInfinite(boolean isInfinite);
